requirements.txt
streamlit
numpy==1.26.4
Pillow
joblib
scikit-learn
tensorflow==2.16.1
deepface
opencv-python-headless==4.11.0.86
